✅ Move signals+store coverage to integration tests
