+++ packages/integration-tests/package.json
@@ -8,6 +8,7 @@
   },
   "devDependencies": {
     "@happy-dom/global-registrator": "^20.0.10",
+    "@preact/signals-core": "^1.14.1",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/react": "^16.3.2",
     "@types/react": "^19.2.14",
@@ -17,6 +18,8 @@
     "@umpire/json": "workspace:^",
     "@umpire/react": "workspace:^",
     "@umpire/reads": "workspace:^",
+    "@umpire/signals": "workspace:^",
+    "@umpire/store": "workspace:^",
     "@umpire/zod": "workspace:^",
     "bun-types": "^1.3.12",
     "react": "^19.2.4",
