+++ packages/integration-tests/tsconfig.json
@@ -15,6 +15,11 @@
       "@umpire/react/*": ["../react/src/*"],
       "@umpire/reads": ["../reads/src/index.ts"],
       "@umpire/reads/*": ["../reads/src/*"],
+      "@umpire/signals": ["../signals/src/index.ts"],
+      "@umpire/signals/preact": ["../signals/src/adapters/preact.ts"],
+      "@umpire/signals/*": ["../signals/src/*"],
+      "@umpire/store": ["../store/src/index.ts"],
+      "@umpire/store/*": ["../store/src/*"],
       "@umpire/zod": ["../zod/src/index.ts"],
       "@umpire/zod/*": ["../zod/src/*"]
     }
